Create findContacts.py
I wanted to get in touch with a bunch of people and gather data about cost of badminton courts around Brisbane so I found a list of halls and email addresses for the organisation. Then decided to auto find their email addresses from their homepage.. 

First check list of home page addresses to see which ones are valid. 
Then look for the word 'contact' in any links on this page. 
Store link. 
Look for email addresses on contact page that contain 'info@', 'enquiries@', 'admin@', 'office@' and store this result. 


Could've cleaned up code to make it look pretty but I actually used this as a chance to store my mistakes so when I go back over this or someone goes through this they see a little bit of how someone arrived at the final solution so that the coder doesn't seem like some genius haha capable of whipping up code but someone who made mistakes and perservered..