Webpage link error fixed! Bloody 'Meta_refresh'...
Had a problem with accessing the library catalogue: 

I thought the webpage for the library catalogue was:   https://elibcat.library.brisbane.qld.gov.au

So I used the homepage but no forms found on this page.. I wondered why. I printed the HTML for the code which was very small.. Which didn't make sense.. 

In the browser this link actually took us elsewhere.. e.g. to https://elibcat.library.brisbane.qld.gov.au/uhtbin/cgisirsi/?ps=ohIqbErG80/ZZELIBCAT/1640104/60/1308/X

In using this link in mechanize it wouldn't redirect (not if I just ran br.open(url)  anyway.. 

So I thought using that long redirect link (that I got from the browser) would fix the problem in our code. 

Worked once or twice? Didn't work after that.. Perhaps because I also searched in my browser on this link and 'used it up'.. So then this code wouldn't run properly - webpage had changed? 

Turns out it's a bit more complicated.. I checked out the HTML for the homepage: https://elibcat.library.brisbane.qld.gov.au

Noticed it was so small! 

<HTML>
<HEAD>
<META HTTP-EQUIV="Refresh" CONTENT="0; URL=/uhtbin/webcat">
</HEAD>
<BODY>
</BODY>
</HTML>

Notice the "URL=/uhtbin/webcat"    I've seen links that start with a '/' before.. You just tack this on to the end of the home page.. 

So I tried adding this on to the end of the home page url - thus giving us: https://elibcat.library.brisbane.qld.gov.au/uhtbin/webcat

This worked in code + in the browser! 

On investigating the meta http-equiv = "refresh" I found this useful link
https://en.wikipedia.org/wiki/Meta_refresh

The below explanation on the wiki page tells us what we need to know: 
"...Used as a redirect..  It is also possible to instruct the browser to fetch a different URL when the page is refreshed, by including the alternative URL in the content parameter. By setting the refresh time interval to zero (or a very low value), this allows meta refresh to be used as a method of URL redirection.

This is exactly the pattern we see in the html for https://elibcat.library.brisbane.qld.gov.au/:
<META HTTP-EQUIV="Refresh" CONTENT="0; URL=/uhtbin/webcat"> 


Sweet. So I now understand this a lot better + things are working!